requirements.txt
python-dotenv==1.1.0
groq==0.25.0
google-generativeai==0.8.5
supabase==2.13.0
PyPDF2==3.0.1
flask==3.1.1
gunicorn==23.0.0
flask-cors==5.0.1